MYR-810: modularised components on MyWallet
