+++ src/components-v2/BalanceDetailList/index.ts
@@ -1,0 +1,4 @@
+import {BalanceDetailList} from './BalanceDetailList';
+import {useStyles} from './balance-detail-list.style';
+
+export {useStyles, BalanceDetailList};
