+++ src/components-v2/MyWallet/index.ts
@@ -1,0 +1,5 @@
+import {MyWallet} from './MyWallet';
+import {BalanceDetail, MyWalletProps} from './my-wallet.interfaces';
+
+export {MyWallet};
+export type {BalanceDetail, MyWalletProps};
