+++ src/components-v2/PrimaryCoinMenu/index.ts
@@ -1,0 +1,4 @@
+import {DraggableBalanceCard, PrimaryCoinMenu} from './PrimaryCoinMenu';
+import {useStyles} from './primary-coin-menu.style';
+
+export {useStyles, DraggableBalanceCard, PrimaryCoinMenu};
